+++ lib/teaspoon/instrumentation.rb
@@ -1,4 +1,5 @@
 require "sprockets/environment"
+require "teaspoon/coverage"
 
 module Teaspoon
   class Instrumentation
@@ -14,7 +15,8 @@ module Teaspoon
         env["QUERY_STRING"].to_s =~ /instrument=(1|true)/ &&            # the instrument param was provided
         response[0] == 200 &&                                           # the status is 200 (304 maybe?)
         response[1]["Content-Type"].to_s == "application/javascript" && # the format is something that we care about
-        response[2].respond_to?(:source)                                # it looks like an asset
+        response[2].respond_to?(:source) &&                             # it looks like an asset
+        !ignored?(response[2])                                          # it is not ignored
     end
 
     def self.executable
@@ -40,6 +42,14 @@ module Teaspoon
     end
 
     protected
+
+    def self.ignored?(asset)
+      Array(Teaspoon::Coverage.configuration.ignore).any? do |ignore|
+        asset.pathname.to_s.match(ignore)
+      end
+    rescue Teaspoon::UnknownCoverage
+      false
+    end
 
     def add_instrumentation(asset)
       source_path = asset.pathname.to_s