+++ lib/teaspoon/suite.rb
@@ -13,7 +13,7 @@ module Teaspoon
     end
 
     attr_accessor :config, :name
-    delegate :helper, :stylesheets, :javascripts, :boot_partial, :body_partial, :no_coverage, :hooks,
+    delegate :helper, :stylesheets, :javascripts, :boot_partial, :body_partial, :hooks,
              to: :config
 
     def initialize(options = {})
@@ -31,21 +31,6 @@ module Teaspoon
       assets = specs
       assets.unshift(helper) if include_helper && helper
       asset_tree(assets)
-    end
-
-    def include_spec?(file)
-      glob.include?(file)
-    end
-
-    def ignored_file?(file)
-      for ignored in no_coverage
-        if ignored.is_a?(String)
-          return true if File.basename(file) == ignored
-        elsif ignored.is_a?(Regexp)
-          return true if file =~ ignored
-        end
-      end
-      false
     end
 
     def include_spec_for?(file)
@@ -82,8 +67,12 @@ module Teaspoon
 
     def instrument_file?(file)
       return false unless @options[:coverage] || Teaspoon.configuration.use_coverage
-      return false if include_spec?(file) || ignored_file?(file)
+      return false if matched_spec_file?(file)
       true
+    end
+
+    def matched_spec_file?(file)
+      glob.include?(file)
     end
 
     def asset_from_file(original)