+++ CHANGELOG.md
@@ -59,7 +59,7 @@
 * Provide QUnit tests with a default `equal` message
 * Support cli reusing running Rails server
 * Add support for Capybara Webkit
-* Better support for RequireJS
+* Better support for RequireJS and more consistent coverage configuration (@davestevens)
 * Add RSpec HTML formatter
 
 #### Bug Fixes